break up `janus_core::task`
Moves handling of authentication tokens to a new module
`janus_core::auth_tokens`. Once that's gone, the only thing of substance
left is `enum VdafInstance` and the macros for dispatching on VDAFs, so
that module is renamed to `janus_core::vdaf` to better describe its
purpose.
